+++ src/mink/limits/__init__.py
@@ -4,6 +4,9 @@ from .collision_avoidance_limit import (
     CollisionAvoidanceLimit as CollisionAvoidanceLimit,
 )
 from .configuration_limit import ConfigurationLimit as ConfigurationLimit
+from .free_joint_velocity_limit import (
+    FreeJointVelocityLimit as FreeJointVelocityLimit,
+)
 from .limit import Constraint as Constraint
 from .limit import Limit as Limit
 from .velocity_limit import VelocityLimit as VelocityLimit